feat(frontend/auth): unify auth success flows and improve forgot/reset UX
redirect successful register/forgot/reset flows to login using query flags
centralize success notices in login (verified/registered/passwordResetRequested/passwordReset)
make forgot-password email validation realtime, matching register behavior
extract ResetPasswordForm as a shared component for forgot/reset flow
harden verify-email flow to reduce false-error flashing during token/query resolution
update forgot-password footer navigation (register + login) and fix mobile form height
clean up and format auth layout/styles
