+++ frontend/app/pages/auth/login/(components)/LoginForm.vue
@@ -1,11 +1,11 @@
 <script setup lang="ts">
 import type { FetchError } from 'ofetch';
+import type { LocationQueryValue } from 'vue-router';
 
 const localePath = useLocalePath();
 const router = useRouter();
 const route = useRoute();
 const loading = ref<boolean>(false);
-const verified = ref(route.query.verified === 'true');
 const useUser = useUserStore();
 const error = reactive({ email: '', password: '', generic: '' });
 const isViewingPassword = ref<boolean>(false);
@@ -25,6 +25,22 @@ const {
   onCredential: handleGoogleCredential,
   buttonText: 'signin_with',
 });
+
+function isTrueQueryFlag(value: LocationQueryValue | LocationQueryValue[] | undefined): boolean {
+  if (!value) return false;
+  const resolvedValue = Array.isArray(value) ? value[0] : value;
+  return resolvedValue === 'true';
+}
+
+function resolveNoticeMessage(): string {
+  if (isTrueQueryFlag(route.query.verified)) return 'auth.verification.success';
+  if (isTrueQueryFlag(route.query.registered)) return 'auth.register.success';
+  if (isTrueQueryFlag(route.query.passwordResetRequested)) return 'auth.password_reset.check_email';
+  if (isTrueQueryFlag(route.query.passwordReset)) return 'auth.password_reset.success';
+  return '';
+}
+
+const noticeMessage = ref<string>(resolveNoticeMessage());
 
 async function handleGoogleCredential(idToken: string) {
   loading.value = true;
@@ -46,7 +62,7 @@ async function handleGoogleCredential(idToken: string) {
 
 async function login() {
   loading.value = true;
-  verified.value = false;
+  noticeMessage.value = '';
   Object.assign(error, { email: '', password: '', generic: '' });
   try {
     await useUser.login(formData.email, formData.password, formData.rememberMe);
@@ -88,12 +104,12 @@ function togglePassword() {
   <div class="form-width">
     <h3 class="text-center text-reactive-primary">{{ $t('auth.login.title') }}</h3>
     <div
-      v-if="verified"
+      v-if="noticeMessage"
       class="alert alert-success d-flex align-items-center mt-2 mb-0"
       role="alert"
     >
       <i class="bi bi-check-circle-fill me-2" />
-      <span>{{ $t('auth.verification.success') }}</span>
+      <span>{{ $t(noticeMessage) }}</span>
     </div>
     <hr class="my-2" />
     <form novalidate>
